Move ietfl3vpnsvc utils to api
