+++ northbound/api/pom.xml
@@ -23,6 +23,16 @@ and is available at http://www.eclipse.org/legal/epl-v10.html
 
     <dependencies>
         <dependency>
+            <groupId>org.opendaylight.ansible</groupId>
+            <artifactId>mdsalutils</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.opendaylight.ansible.model</groupId>
+            <artifactId>ietf-l3vpn-svc</artifactId>
+            <version>2017-05-02.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>org.opendaylight.mdsal.model</groupId>
             <artifactId>ietf-topology</artifactId>
         </dependency>
